Use long for point balances

+++ src/main/java/com/gchoi/loyalty/dto/BalanceResponse.java
@@ -5,5 +5,5 @@ import com.gchoi.loyalty.entity.Tier;
 /**
  * Response returned when a customer checks their available points and current tier.
  */
-public record BalanceResponse(String customerId, int availablePoints, Tier tier) {
+public record BalanceResponse(String customerId, long availablePoints, Tier tier) {
 }
